backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
python-dotenv
httpx
weasyprint
PyMuPDF
python-docx
pandas
openpyxl
Pillow
jinja2
groq
